Se implemento pause_ticks y el blink de bash
Tambien se creo un archivo propio para printmem para aislar sus funciones propias del resto de programas

+++ Kernel/drivers/time.c
@@ -1,17 +1,17 @@
 #include <time.h>
 
-static unsigned long ticks = 0;
+static uint64_t ticks = 0;
 
-void timer_handler() {							// Es la funcion que se ejecuta cuando ocurra la interrupcion del timer tick (Incrementa la cantidad de ticks)
+void timer_handler() {							    // Es la funcion que se ejecuta cuando ocurra la interrupcion del timer tick (Incrementa la cantidad de ticks)
 	ticks++;
     //if(ticks%2==0)
         change_context();
 }
 
-int ticks_elapsed() {							// Devuelve la cantidad de veces que se ejecuto la interrupcion del timer tick desde que se inicio el sistema
+uint64_t ticks_elapsed() {							// Devuelve la cantidad de veces que se ejecuto la interrupcion del timer tick desde que se inicio el sistema
 	return ticks;
 }
 
-int seconds_elapsed() {							// Devuelve la cantidad de segundos que transcurrieron desde que se inicio el sistema (Se realiza un tick cada 18,2 segs)
+uint64_t seconds_elapsed() {					    // Devuelve la cantidad de segundos que transcurrieron desde que se inicio el sistema (Se realiza un tick cada 18,2 segs)
 	return ticks / 18;
 }
